Update Lens framework
