+++ Demo/EmptyModule/EmptyModuleViewController.swift
@@ -7,21 +7,49 @@
 //
 
 import HHModule
+#if HHLens
+import HHLens
+#endif
 
 final class EmptyModuleViewController: ARCHViewController<EmptyModuleState, EmptyModuleEventHandler> {
 
-    let label = Label()
+    private let label = Label()
+    private let button = UIButton()
 
     // MARK: - View life cycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        label.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        label.frame = view.bounds
-        label.textAlignment = .center
-        view.addSubview(label)
-
         view.backgroundColor = .white
+
+        button.setTitle("BUTTON", for: .normal)
+        label.text = "LABEL"
+
+        [button, label].forEach {
+            $0.translatesAutoresizingMaskIntoConstraints = false
+            self.view.addSubview($0)
+        }
+
+        NSLayoutConstraint.activate([
+            button.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 15.0),
+            button.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -15.0),
+            button.centerYAnchor.constraint(equalTo: view.centerYAnchor),
+
+            label.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 15.0),
+            label.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -15.0),
+            label.bottomAnchor.constraint(equalTo: button.topAnchor)
+        ])
+
+        configure()
+    }
+
+    // MARK: - Configure
+
+    private func configure() {
+#if HHLens
+        _ = button |> kRoundedGreyButtonStyle
+        _ = label |> kGreenLabelStyle
+#endif
     }
 }
