+++ Source/HHLens/PreludeUI/Lens/CALayerLenses.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public protocol CALayerProtocol: KSObjectProtocol {
@@ -22,7 +21,7 @@ extension LensHolder where Object: CALayerProtocol {
             set: { $1.borderColor = $0; return $1 }
         )
     }
-    
+
     public var borderWidth: Lens<Object, CGFloat> {
         return Lens(
             view: { $0.borderWidth },
@@ -35,42 +34,42 @@ extension LensHolder where Object: CALayerProtocol {
             set: { $1.cornerRadius = $0; return $1 }
         )
     }
-    
+
     public var masksToBounds: Lens<Object, Bool> {
         return Lens(
             view: { $0.masksToBounds },
             set: { $1.masksToBounds = $0; return $1 }
         )
     }
-    
+
     public var shadowColor: Lens<Object, CGColor?> {
         return Lens(
             view: { $0.shadowColor },
             set: { $1.shadowColor = $0; return $1 }
         )
     }
-    
+
     public var shadowOffset: Lens<Object, CGSize> {
         return Lens(
             view: { $0.shadowOffset },
             set: { $1.shadowOffset = $0; return $1 }
         )
     }
-    
+
     public var shadowOpacity: Lens<Object, Float> {
         return Lens(
             view: { $0.shadowOpacity },
             set: { $1.shadowOpacity = $0; return $1 }
         )
     }
-    
+
     public var shadowRadius: Lens<Object, CGFloat> {
         return Lens(
             view: { $0.shadowRadius },
             set: { $1.shadowRadius = $0; return $1 }
         )
     }
-    
+
     public var shouldRasterize: Lens<Object, Bool> {
         return Lens(
             view: { $0.shouldRasterize },
