+++ Source/HHLens/PreludeUI/Lens/CGPointLenses.swift
@@ -8,7 +8,7 @@ extension CGPoint {
             view: { $0.x },
             set: { .init(x: $0, y: $1.y) }
         )
-        
+
         public static let y = Lens<CGPoint, CGFloat>(
             view: { $0.y },
             set: { .init(x: $1.x, y: $0) }
