+++ Source/HHLens/PreludeUI/Lens/CGSizeLenses.swift
@@ -8,7 +8,7 @@ extension CGSize {
             view: { $0.width },
             set: { .init(width: $0, height: $1.height) }
         )
-        
+
         public static let height = Lens<CGSize, CGFloat>(
             view: { $0.height },
             set: { .init(width: $1.width, height: $0) }
