+++ Source/HHLens/PreludeUI/Lens/NSObject.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 extension NSObject: LensObject {}
