+++ Source/HHLens/PreludeUI/Lens/UIControlLenses.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public typealias TargetSelectorControlEvent = (Any, Selector, UIControl.Event)
@@ -21,7 +20,7 @@ public protocol UIControlProtocol: UIViewProtocol {
 extension UIControl: UIControlProtocol {}
 
 public extension LensHolder where Object: UIControlProtocol {
-    
+
     public var targets: Lens<Object, [TargetSelectorControlEvent]> {
         return Lens(
             view: allTargetsSelectorsAndEvents(for:),
@@ -29,38 +28,38 @@ public extension LensHolder where Object: UIControlProtocol {
                 control.removeTarget(nil, action: nil, for: .allEvents)
                 targets.forEach(control.addTarget(_:action:for:))
                 return control
-        }
+            }
         )
     }
-    
+
     public var contentHorizontalAlignment: Lens<Object, UIControl.ContentHorizontalAlignment> {
         return Lens(
             view: { $0.contentHorizontalAlignment },
             set: { $1.contentHorizontalAlignment = $0; return $1 }
         )
     }
-    
+
     public var contentVerticalAlignment: Lens<Object, UIControl.ContentVerticalAlignment> {
         return Lens(
             view: { $0.contentVerticalAlignment },
             set: { $1.contentVerticalAlignment = $0; return $1 }
         )
     }
-    
+
     public var isEnabled: Lens<Object, Bool> {
         return Lens(
             view: { $0.isEnabled },
             set: { $1.isEnabled = $0; return $1 }
         )
     }
-    
+
     public var isHighlighted: Lens<Object, Bool> {
         return Lens(
             view: { $0.isHighlighted },
             set: { $1.isHighlighted = $0; return $1 }
         )
     }
-    
+
     public var isSelected: Lens<Object, Bool> {
         return Lens(
             view: { $0.isSelected },
@@ -78,7 +77,7 @@ public extension LensHolder where Object: UIControlProtocol {
  */
 private func allTargetsSelectorsAndEvents(for control: UIControlProtocol)
     -> [TargetSelectorControlEvent] {
-        
+
         return bitComponents(forMask: control.allControlEvents.rawValue)
             .map(UIControl.Event.init(rawValue:))
             .flatMap { event in
@@ -86,10 +85,10 @@ private func allTargetsSelectorsAndEvents(for control: UIControlProtocol)
                     .flatMap { target in
                         (control.actions(forTarget: target, forControlEvent: event) ?? [])
                             .map { action in
-                                return (target as Any, Selector(action), event)
-                        }
-                }
-        }
+                                (target as Any, Selector(action), event)
+                            }
+                    }
+            }
 }
 
 func bitComponents(forMask mask: UInt) -> [UInt] {
