+++ Source/HHLens/PreludeUI/Lens/UIImageViewLenses.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public protocol UIImageViewProtocol: UIViewProtocol {
