+++ Source/HHLens/PreludeUI/Lens/UILabelLenses.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public protocol UILabelProtocol: UIViewProtocol {
@@ -22,56 +21,56 @@ public extension LensHolder where Object: UILabelProtocol {
             set: { $1.adjustsFontSizeToFitWidth = $0; return $1; }
         )
     }
-    
+
     public var attributedText: Lens<Object, NSAttributedString?> {
         return Lens(
             view: { $0.attributedText },
             set: { $1.attributedText = $0; return $1; }
         )
     }
-    
+
     public var font: Lens<Object, UIFont> {
         return Lens(
             view: { $0.font },
             set: { $1.font = $0; return $1; }
         )
     }
-    
+
     public var lineBreakMode: Lens<Object, NSLineBreakMode> {
         return Lens(
             view: { $0.lineBreakMode },
             set: { $1.lineBreakMode = $0; return $1; }
         )
     }
-    
+
     public var minimumScaleFactor: Lens<Object, CGFloat> {
         return Lens(
             view: { $0.minimumScaleFactor },
             set: { $1.minimumScaleFactor = $0; return $1; }
         )
     }
-    
+
     public var numberOfLines: Lens<Object, Int> {
         return Lens(
             view: { $0.numberOfLines },
             set: { $1.numberOfLines = $0; return $1; }
         )
     }
-    
+
     public var text: Lens<Object, String> {
         return Lens(
             view: { $0.text ?? "" },
             set: { $1.text = $0; return $1; }
         )
     }
-    
+
     public var textAlignment: Lens<Object, NSTextAlignment> {
         return Lens(
             view: { $0.textAlignment },
             set: { $1.textAlignment = $0; return $1; }
         )
     }
-    
+
     public var textColor: Lens<Object, UIColor> {
         return Lens(
             view: { $0.textColor },
