+++ Source/HHLens/PreludeUI/Lens/UITextFieldLenses.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public protocol UITextFieldProtocol: UIControlProtocol, UITextInputTraitsProtocol {
@@ -13,42 +12,42 @@ public protocol UITextFieldProtocol: UIControlProtocol, UITextInputTraitsProtoco
 extension UITextField: UITextFieldProtocol {}
 
 public extension LensHolder where Object: UITextFieldProtocol {
-    
+
     public var borderStyle: Lens<Object, UITextField.BorderStyle> {
         return Lens(
             view: { $0.borderStyle },
             set: { $1.borderStyle = $0; return $1 }
         )
     }
-    
+
     public var font: Lens<Object, UIFont?> {
         return Lens(
             view: { $0.font },
             set: { $1.font = $0; return $1 }
         )
     }
-    
+
     public var placeholder: Lens<Object, String?> {
         return Lens(
             view: { $0.placeholder },
             set: { $1.placeholder = $0; return $1 }
         )
     }
-    
+
     public var textAlignment: Lens<Object, NSTextAlignment> {
         return Lens(
             view: { $0.textAlignment },
             set: { $1.textAlignment = $0; return $1 }
         )
     }
-    
+
     public var textColor: Lens<Object, UIColor?> {
         return Lens(
             view: { $0.textColor },
             set: { $1.textColor = $0; return $1 }
         )
     }
-    
+
     public var text: Lens<Object, String?> {
         return Lens(
             view: { $0.text },
