+++ Source/HHLens/PreludeUI/Lens/UITraitEnvironment.swift
@@ -1,4 +1,3 @@
-
 import UIKit
 
 public protocol UITraitEnvironmentProtocol: NSObjectProtocol {
